[spark] Rename BatteryTestRegression to BatterySuccessPredictor
